Add files via upload
This is the actual code that will be run on the MKR that is to be inserted into the PCBs.
it currently has a few bugs and variable names aren't great since its basically an extended version of a pushingbox API test that was done to make sure data could be sent to a google sheets document. i haven't changed all the names yet since it means tweaking them in google sheets, pushingbox, and the code which is a lot of places where things can go wrong. fixing variable names/API names will be one of the last steps i plan to take.