better names, first in a series of updates
after a series of failures i have began to take a more systematic approach to fixing the naming of the stuff that allows the MKR to talk to pushingbox and google spreadsheets. step 1 was updating the variable names being used in the arduino sketch to be more human readable. this code has been tested and absolutely sends the right data to the right spreadsheet (yay!)